+++ tests/unit/models/reddit/test_widgets.py
@@ -1,0 +1,15 @@
+from praw.models import (SubredditWidgets, SubredditWidgetsModeration,
+                         Widget, WidgetModeration)
+
+from ... import UnitTest
+
+
+class TestWidgets(UnitTest):
+    def test_subredditwidgets_mod(self):
+        sw = SubredditWidgets(self.reddit.subreddit('fake_subreddit'))
+        assert isinstance(sw.mod, SubredditWidgetsModeration)
+
+    def test_widget_mod(self):
+        w = Widget(self.reddit, {})
+        assert isinstance(w.mod, WidgetModeration)
+        assert w.mod.widget == w