+++ CHANGES.rst
@@ -9,6 +9,12 @@ Unreleased
 * Add method :meth:`.Redditor.trophies` to get a list of the Redditor's
   trophies.
 * Add class :class:`.PostFlairWidget`.
+* Add class :class:`.SubredditWidgetsModeration` (accessible through
+  :attr:`.SubredditWidgets.mod`) and method :meth:`.add_text_area`.
+* Add class :class:`.WidgetModeration` (accessible through the ``.mod``
+  attribute on any widget) with methods :meth:`~.WidgetModeration.update` and
+  :meth:`~.WidgetModeration.delete`.
+* Add method :meth:`.Reddit.put` for HTTP PUT requests.
 
 **Changed**